Module Gedcom.GedcomHelpers

val concat : gedcom_line list -> gedcom_line list

Concatenation of CONT / CONC lines groups.

val trim_xref : string -> string

Trim characters which are not part of the xref (i.e. leading and trailing ' ' and '@')

val parse_name : string -> string option * string option * string option

Split name in first name / last name / title.

val name_first : (string option * string option * string option) -> string

Extract first name part of a gedcom name or raise Not_found.

val name_last : (string option * string option * string option) -> string

Extract last name part of a gedcom name or raise Not_found.

val name_title : (string option * string option * string option) -> string

Extract title part of a gedcom name or raise Not_found.